refactor(agents): extract resolveBin + withRetry into shared lib
The PATH-ENOENT fix (#74) and pooler-retry fix (#73) were both generic but copied
per-agent. Consolidate into scripts/lib/agent-resilience.mjs:

- resolveBin(name, candidates): minimal-PATH-robust binary resolution (env override ->
  validated `which` -> known install locations -> bare name).
- withRetry(fn, label) + isTransientDbError(message): retry transient pooler/connection
  failures (network throws AND transient error values), pass real errors through.

Migrations:
- trust-remediation-loop, refresh-youth-justice-source-chain: use shared resolveBin.
- run-bridge-fuzzy: NEW FIX -- it spawned bare `psql` with no resolution, so it had the
  same ENOENT bug under cron/pm2's minimal PATH; now resolves via the shared helper.
- sync-tracker-evidence, run-tracker-refresh: use shared withRetry/isTransientDbError.

Behaviour-neutral consolidation (the shared functions are identical to the inline ones)
plus the run-bridge-fuzzy hardening. Verified: lib smoke test under a minimal PATH
(resolveBin -> /opt/homebrew/bin/psql; withRetry recovers a transient 'schema cache'),
and sync-tracker-evidence runs clean end-to-end via the imported helper (39 rows, 25s).

Follow-up: trust-remediation-loop has its own older withRetry(label, fn) with a different
signature; left as-is to avoid churn, worth unifying later.
